- Criação do metodo que retorna a reserva que o adm Clicou na tela de listar.

+++ app/src/main/java/com/example/alexandrecardoso/projetohotelfei/Avaliacao.java
@@ -2,21 +2,21 @@ package com.example.alexandrecardoso.projetohotelfei;
 
 public class Avaliacao {
 
-    private double nota;
+    private float nota;
     private String titulo;
     private String mensagem;
 
-    public Avaliacao(String titulo, double nota, String mensagem){
+    public Avaliacao(String titulo, float nota, String mensagem){
         this.nota = nota;
         this.mensagem = mensagem;
         this.titulo = titulo;
     }
 
-    public double getNota() {
+    public float getNota() {
         return nota;
     }
 
-    public void setNota(double nota) {
+    public void setNota(float nota) {
         this.nota = nota;
     }
